# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "电商订单支付系统 API 文档",
        "title": "Order Payment System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/home/order": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "创建订单",
                "parameters": [
                    {
                        "description": "订单信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_internal_types.OrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                        }
                    }
                }
            }
        },
        "/home/order/cancel": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "取消订单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单号",
                        "name": "orderNo",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                        }
                    }
                }
            }
        },
        "/home/order/cart_checkout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "购物车结算",
                "parameters": [
                    {
                        "description": "结算信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_internal_types.CartCheckoutReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/home/order/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "完成订单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单号",
                        "name": "order_no",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                        }
                    }
                }
            }
        },
        "/home/order/confirm": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "确认收货",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单号",
                        "name": "order_no",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                        }
                    }
                }
            }
        },
        "/home/order/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "订单日志",
                "parameters": [
                    {
                        "type": "string",
                        "description": "订单号",
                        "name": "order_no",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/order-payment-system_internal_types.OrderLogResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/home/order/ship": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "商家发货",
                "parameters": [
                    {
                        "description": "物流信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_internal_types.ShipRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                        }
                    }
                }
            }
        },
...
        "/home/search/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "订单"
                ],
                "summary": "用户订单列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/order-payment-system_pkg_response.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/order-payment-system_internal_types.OrderResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "order-payment-system_internal_types.AddressResp": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_default": {
                    "type": "boolean"
                },
                "phone": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "receiver": {
                    "type": "string"
                }
            }
        },
        "order-payment-system_internal_types.CartCheckoutReq": {
            "type": "object",
            "required": [
                "address_id"
            ],
            "properties": {
                "address_id": {
                    "description": "必填",
                    "type": "integer"
                }
            }
        },
...
        "order-payment-system_internal_types.OrderItemResp": {
            "type": "object",
            "properties": {
                "goods_id": {
                    "type": "integer"
                },
                "goods_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "order-payment-system_internal_types.OrderLogResp": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "new_status": {
                    "type": "integer"
                },
                "old_status": {
                    "type": "integer"
                },
                "order_no": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
        "order-payment-system_internal_types.OrderRequest": {
            "type": "object",
            "properties": {
                "address_id": {
                    "description": "收货地址ID",
                    "type": "integer"
                },
                "buyNum": {
                    "type": "integer"
                },
                "goodsId": {
                    "type": "integer"
                }
            }
        },
        "order-payment-system_internal_types.OrderResp": {
            "type": "object",
            "properties": {
                "address_snapshot": {
                    "$ref": "#/definitions/order-payment-system_internal_types.AddressResp"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/order-payment-system_internal_types.OrderItemResp"
                    }
                },
                "order_no": {
                    "type": "string"
                },
                "pay_time": {
                    "type": "string"
                },
                "ship_company": {
                    "type": "string"
                },
                "ship_no": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "status_name": {
                    "type": "string"
                },
                "total_price": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "order-payment-system_internal_types.ShipRequest": {
            "type": "object",
            "required": [
                "order_no",
                "ship_company",
                "ship_no"
            ],
            "properties": {
                "order_no": {
                    "type": "string"
                },
                "ship_company": {
                    "type": "string"
                },
                "ship_no": {
                    "type": "string"
                }
            }
        },
...
        "order-payment-system_pkg_response.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据"
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}